Add maximum opacity setting for fadejs w/ html tag attr
Currently, when using the fadejs functionality, if the target element also has a stylesheet class that sets it's opacity below 100%, the fadejs rendering will make the element appear as if it were set to 100%

Added logic that can extract a custom html tag attribute, "fadejs-data", that a page can use to define an elements maximum opacity. This will allow the fadejs rendered elements to match the intended opacity for the element.

The opacity property of the fadejs-data attribute takes a value between 0-1, and the regex query used to select the string supports whether a trailing semicolon follows the value or not. Uses 1, by default (if the attribute is missing from the tag), which emulates the same behavior as fadejs had before.